+++ SysconCommon/SMBClasses/apttyp.cs
@@ -2,10 +2,16 @@ using System;
 
 namespace SMB.Tables {
 	public class apttyp : smbtable {
+		/// <summary>
+		/// Appointment#                    
+		/// </summary>
 		[SMBFieldName("recnum")]
 		[SMBFieldDescription("Appointment#")]
-		public int recnum;
+		public long recnum;
 
+		/// <summary>
+		/// Description                     
+		/// </summary>
 		[SMBFieldName("dscrpt")]
 		[SMBFieldDescription("Description")]
 		public string dscrpt;
