+++ SysconCommon/SMBClasses/asmcls.cs
@@ -2,25 +2,40 @@ using System;
 
 namespace SMB.Tables {
 	public class asmcls : smbtable {
+		/// <summary>
+		/// Class#                          
+		/// </summary>
 		[SMBFieldName("recnum")]
 		[SMBFieldDescription("Class#")]
-		public int recnum;
+		public long recnum;
 
+		/// <summary>
+		/// Class Name                      
+		/// </summary>
 		[SMBFieldName("clsnme")]
 		[SMBFieldDescription("Class Name")]
 		public string clsnme;
 
+		/// <summary>
+		/// Indention Level                 
+		/// </summary>
 		[SMBFieldName("indent")]
 		[SMBFieldDescription("Indention Level")]
-		public int indent;
+		public long indent;
 
+		/// <summary>
+		/// Class Parent#                   
+		/// </summary>
 		[SMBFieldName("parcls")]
 		[SMBFieldDescription("Class Parent#")]
-		public int parcls;
+		public long parcls;
 
+		/// <summary>
+		/// Class Has Children              
+		/// </summary>
 		[SMBFieldName("haskid")]
 		[SMBFieldDescription("Class Has Children")]
-		public int haskid;
+		public long haskid;
 
 	}
 }
